requirements.txt
httpx
google-search-results
dnspython
